Piecing apart a previous PR to try and sort out failing regression tests
