+++ cd/application-deployment/prod/vaec-celery-task-definition.json
@@ -192,14 +192,6 @@
         {
           "name": "COMP_AND_PEN_DYNAMODB_NAME",
           "value": "prod-bip-payment-notification-table"
-        },
-        {
-          "name": "VA_PROFILE_V3_COMBINE_CONTACT_INFO_AND_PERMISSIONS_LOOKUP",
-          "value": "True"
-        },
-        {
-          "name": "VA_PROFILE_V3_IDENTIFY_MOBILE_TELEPHONE_NUMBERS",
-          "value": "True"
         }
       ],
       "secrets": [
